Remove orphan `Show` instance for `Foreign.C.Error.Errno`
This orphan instance largely exists to auto-derive a `Show` instance for
`FsError`. Exposing an orphan instance for a type from `base` is not very nice,
however. As such, we've removed the orphan instance, and we've written the `Show
FsError` instance by hand.

Note that this is a breaking change for both `fs-api` and `fs-sim`, since the
orphan instance is propagated from `fs-api` through `fs-sim`.