Comparative benchmarks for `hGetSome(at)` and their replacement functions.
`hGetSome'` is a new function that provides the same functionality as
`hGetSome`, but uses the `hGetSomeBuf` primitive. Similary, we implement a new
`hGetSomeAt'` function that provides the same functionality as `hGetSomeAt`.
These comparative benchmarks should show whether we can replace the
`hGetSome(At)` primitives with `hGetBufSome(At)` primtives and the new compound
functions.