Make `Stream`s definitely finite or definitely infinite
Where streams could previously be *definitely* finite or *possibly* infinite,
they should now be *definitely* finite or *definitely* infinite. This is mostly
a conceptual change: it was already guaranteed by most if not all of the
`Stream` functions. Still, the conceptual change should make the use of
`Streams` more ergonomic going forward.

As a result of and in addition to the conceptual change, the `Stream`
interface got an overhaul. The concrete changes are listed in the changelog.

Finally, we add a bunch of tests for the `Stream` module.